refactor: 영역 별로 폴더 분리하기
