+++ src/refactoring/components/AdminPage/couponSection/CouponForm.tsx
@@ -1,7 +1,7 @@
 import { useState, ChangeEvent } from 'react';
 
-import { Coupon } from '../../../types';
-import { DEFAULT_COUPON } from '../../constants';
+import type { Coupon } from '../../../../types';
+import { DEFAULT_COUPON } from '../../../constants';
 
 interface CouponFormProps {
   onSubmit: (coupon: Coupon) => void;
